Move unwindset.{h,cpp} to goto-programs
This code is not specific to goto-instrument and used more widely across
the code base. Removes a now-unnecessary dependency of goto-checker on
goto-instrument.
