+++ src/goto-checker/module_dependencies.txt
@@ -1,7 +1,6 @@
 assembler
 cbmc # symex_bmc will be moved next
 goto-checker
-goto-instrument
 goto-programs
 goto-symex
 langapi
