+++ src/goto-checker/single_loop_incremental_symex_checker.h
@@ -14,9 +14,9 @@ Author: Daniel Kroening, Peter Schrammel
 #ifndef CPROVER_GOTO_CHECKER_SINGLE_LOOP_INCREMENTAL_SYMEX_CHECKER_H
 #define CPROVER_GOTO_CHECKER_SINGLE_LOOP_INCREMENTAL_SYMEX_CHECKER_H
 
-#include <goto-symex/path_storage.h>
+#include <goto-programs/unwindset.h>
 
-#include <goto-instrument/unwindset.h>
+#include <goto-symex/path_storage.h>
 
 #include "goto_symex_property_decider.h"
 #include "goto_trace_provider.h"
